+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -11,7 +11,7 @@ import java.util.stream.Collectors;
 
 @Data
 public class User {
-    private int id;
+    private long id;
 
     private String name;
     @Pattern(regexp = "^.+@.+\\..+$", message = "Некорректный адрес электронной почты")
@@ -21,7 +21,7 @@ public class User {
     private final String login;
     @Birthday
     private final LocalDate birthday;
-    private Set<Integer> friends;
+    private Set<Long> friends;
 
     public User(String name, String email, String login, LocalDate birthday) {
         this.name = name;
@@ -31,21 +31,21 @@ public class User {
         friends = new HashSet<>();
     }
 
-    public void addFriend(int friendId) {
+    public void addFriend(long friendId) {
         friends.add(friendId);
     }
 
-    public void deleteFriend(int friendId) {
+    public void deleteFriend(long friendId) {
         friends.remove(friendId);
     }
 
-    public List<Integer> getCommonFriends(User otherUser) {
-        List<Integer> commonFriendsId;
+    public List<Long> getCommonFriends(User otherUser) {
+        List<Long> commonFriendsId;
         if (this.getId() == otherUser.getId()) {
             commonFriendsId = this.getFriends().stream().collect(Collectors.toList());
             return Collections.unmodifiableList(commonFriendsId);
         }
-        Set<Integer> intersection = new HashSet<>(getFriends());
+        Set<Long> intersection = new HashSet<>(getFriends());
         intersection.retainAll(otherUser.getFriends());
         commonFriendsId = intersection.stream().collect(Collectors.toList());
         return Collections.unmodifiableList(commonFriendsId);
